Move targeted integration tests into their own package
